+++ fixtures/testscript/testscript-example-2.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestScript
+ xmlns="http://hl7.org/fhir">
+    	<id value="TS01"/>
+    <name value="Test Script 1"/>
+    <description value="Example Test Script"/>
+    <fixture
+ id="F1"
+>
+    <uri value="patient/patient-example.xml"/>
+</fixture>
+    <setup
+>
+    <operation
+>
+    <type value="create"/>
+    <source value="F1"/>
+</operation>
+</setup>
+    <test
+ id="READ01"
+>
+    <name value="Read Patient"/>
+    <description value="Read a patient and validate response."/>
+    <metadata
+>
+    <link
+>
+    <url value="http://fhir.org/spec.html"/>
+    <description value="link to the FHIR specification"/>
+</link>
+    <requires
+>
+    <resourceType value="Patient"/>
+    <operations value="create"/>
+</requires>
+    <validates
+>
+    <resourceType value="Patient"/>
+    <operations value="read"/>
+</validates>
+</metadata>
+    <operation
+>
+    <type value="read"/>
+    <target value="F1"/>
+</operation>
+    <assertion
+>
+    <type value="responseOkay"/>
+</assertion>
+<assertion
+>
+    <type value="valid_content_location_present"/>
+</assertion>
+<assertion
+>
+    <type value="Patient"/>
+</assertion>
+</test>
+    <teardown
+>
+    <operation
+>
+    <type value="delete"/>
+    <target value="F1"/>
+</operation>
+</teardown>
+</TestScript>
