+++ lib/plan_executor.rb
@@ -10,7 +10,7 @@ require_relative File.join('.','executor.rb')
 require_relative File.join('.','test_result.rb')
 require_relative File.join('.','resource_generator.rb')
 require_relative File.join('tests','assertions.rb')
-require_relative File.join('tests','base_test.rb')
+require_relative File.join('tests','suites','base_test.rb')
 
 root = File.expand_path '..', File.dirname(File.absolute_path(__FILE__))
 Dir.glob(File.join(root, 'lib','tests','**','*.rb')).each do |file|
