+++ lib/tests/suites/base_test.rb
