+++ lib/tests/suites/multiserver/fetch_patient_record_test.rb
