+++ lib/tests/suites/read_test.rb
