+++ lib/tests/suites/resource_test.rb
