+++ lib/tests/suites/search_test.rb
