+++ lib/tests/suites/sprinkler_search_test.rb
