+++ lib/tests/testscripts/testscript_engine.rb
@@ -1,0 +1,19 @@
+module Crucible
+  module Tests
+    class TestScriptEngine
+
+      def self.list_all
+        {}
+      end
+
+      def list_all_with_conformance(multiserver=false, metadata=nil)
+      	{}
+      end
+
+      def self.tests
+        []
+      end
+
+    end
+  end
+end
